Split works into modules
